NodeGroupScaler
* Allow provisioning of nodes with a predefined state
* Add provisioning of active nodes at first node group discovery

+++ src/node_groups/mod.rs
@@ -17,8 +17,8 @@ pub struct NodeGroup {
 pub struct Config {
     node_bandwidth_capacity: BandwidthCapacity,
     bandwidth_thresholds: BandwidthThresholds,
-    min_nodes: Option<u64>,
-    max_nodes: Option<u64>,
+    min_nodes: Option<u32>,
+    max_nodes: Option<u32>,
 }
 
 #[derive(Debug, Clone, Deserialize)]
